feat: Support Unity6000 (6000.0.46f1)
