+++ Assets/RemoveDevWatermark/package.json
@@ -1,7 +1,7 @@
 {
     "name": "dev.kyubuns.removedevwatermark",
     "displayName": "RemoveDevWatermark",
-    "version": "0.1.1",
+    "version": "1.0.0",
     "unity": "2021.3",
     "description": "Remove Development Build Watermark",
     "documentationUrl": "https://github.com/kyubuns/UnityRemoveDevWatermark",
